# api-docs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Time Here Now API",
    "version": "20251023",
    "description": "API to get the current time based on timezone or client IP. **IMPORTANT: All time values are sourced from NEAR blockchain, NOT system/NTP time.** The API polls NEAR blockchain at 5 Hz (200ms intervals) and serves cached blockchain timestamps. Time zones and offsets are sourced from the IANA Time Zone Database (tzdb). Returns HTTP 503 if blockchain time is unavailable."
  },
  "servers": [
    { "url": "https://timeherenow.rodit.org:8443/api" },
    { "url": "http://localhost:8080/api" }
  ],
...
  "paths": {
...
    "/timers/schedule": {
      "post": {
        "summary": "schedule a delayed webhook using SDK-configured destination.",
        "description": "Schedules a delayed webhook that fires after the specified delay. **All timestamps (scheduled_at, execute_at, fired_at) use NEAR blockchain time, not system time.** The webhook is sent to the destination configured in the RODiT SDK client. The timer uses blockchain time for both scheduling and firing to ensure tamper-proof timing.\n\n**Blockchain Time Granularity:** NEAR blockchain produces blocks at ~500-600ms intervals. Timestamps advance in discrete steps, not continuously. fired_at is guaranteed to be >= execute_at to maintain temporal consistency. Due to blockchain time granularity and caching, fired_at may equal execute_at even when actual delivery occurs later.\n\n**Minimum Granularity:** 1 second. Sub-second delays are not supported due to blockchain time characteristics.\n\n**Persistence:** Timers are automatically saved to disk every hour and restored on server restart. Timers that should have already fired are never sent late - they are skipped on restore. Maximum data loss window is 1 hour between auto-saves.",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["delay_seconds"],
                "properties": {
                  "delay_seconds": { "type": "number", "minimum": 1, "maximum": 172800, "description": "Delay in seconds before webhook is sent (1 second to 48 hours)" },
                  "payload": { "type": "object", "description": "Optional payload echoed back in the webhook" }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Timer scheduled",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TimerScheduleResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/ErrorJsonResponse" },
          "400": { "$ref": "#/components/responses/ErrorJsonResponse" },
          "default": { "$ref": "#/components/responses/ErrorJsonResponse" }
        }
      }
    },
...
  },
  "components": {
    "responses": {
...
      "ErrorJsonResponse": {
        "description": "an error response in JSON format",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorJsonResponse"
            }
          }
        }
      }
    },
    "schemas": {
...
      "ErrorJsonResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "details about the error encountered"
          }
        }
      },
...
      "TimerScheduleResponse": {
        "type": "object",
        "required": ["timer_id", "delay_seconds", "scheduled_at", "execute_at"],
        "properties": {
          "timer_id": { "type": "string", "description": "ULID identifying the scheduled timer" },
          "delay_seconds": { "type": "number", "description": "Delay in seconds before webhook fires" },
          "scheduled_at": { "type": "string", "description": "ISO 8601 timestamp when scheduling occurred (NEAR blockchain time)" },
          "execute_at": { "type": "string", "description": "ISO 8601 timestamp when webhook is scheduled to fire (NEAR blockchain time)" },
          "requestId": { "type": "string", "description": "Server request correlation ID" }
        },
        "description": "Response when scheduling a timer. The webhook payload will include these fields plus 'fired_at' (actual firing time from blockchain) and 'payload' (echoed from request)."
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}